Update the service catalog when the environment changes (Issue #285, PR #287)
# Description

I tried adding a test for this in Cypress, but this was too difficult.
When mocking api calls, I did not find a way to distinguish between 2 service_catalog calls.
Testing this in jest also seem a bit of work because we have to mount the entire application, which means also mocking a bunch of calls.

I added a todo to #286 for writing a test for this scenario.

closes #285

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Code is clear and sufficiently documented
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design