Update design of not found page (Issue #1267, PR #1549)
# Description

Still a draft.
The way we calculate the base url for proxy support is in conflict with the handling of not existing pages.
Urls that are not found as a valid path lead to the 404 page on the one hand, and are understood as the `baseurl` and used in the navigation links on the other hand.
So not only can you end up on a page that doesn't exist, the links to navigate away from there won't work either. (Only the back button and changing the url manually)
I think we should discuss this (and possibly restrict the supported proxy paths to end with `console/`, as they probably already do)

closes #1267

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Code is clear and sufficiently documented
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [x] ~End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design