Refactor Query and Command definitions (Issue #1414, PR #1697)
# Description

I was not able to remove the concepts of queries and commands.
There is too much difference for now to unite them properly.
To not introduce a new concept, I dropped the usage of operations.
I have created followup tickets for the refactors that were started in this PR.

- Remove Poster for CreateCallback
- Remove Deleter for GetCallbacks
- Remove Patcher for InstanceUpdate
- Restructure Query & Command files
- Give consistent names to all page components

closes #1414

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Code is clear and sufficiently documented
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design