Update to version 6 of react-router (Issue #1869, PR #1879)
# Description

This PR has 2 parts. The first part is using the new api provided by react-router v6.
All of the routing logic changes are related to this.
It's mostly replacing references to history with location and navigate.
Typing for routing has also changed a bit (RouteMatch, useRouteParams).
The second part is changing the way an environment change is handled.
Most of the changes are related to removing the environment from the constructor.
And making sure the environmentHandler has an environment to provide to the QueryManagers.

closes #1869

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Code is clear and sufficiently documented
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design