Fix broken styling (Issue #1892, PR #2153)
# Description

The problem was not related to dependency updates, but to the reordering of the imports. The custom styling of the components were overwritten by the `base` styling from Patternfly. Importing the `base` from further up the tree seems to have fixed the issue.

closes #1892

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Code is clear and sufficiently documented
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [x] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design