Document environment data usage (Issue #2172, PR #2353)
# Description

I found that the latest status of this data handling is generally ok (there have been changes since the ticket was created though, so it might be thanks to that).
I've included the analysis nevertheless. If we decide to do one of the suggested improvements, I'll create a separate ticket for that and remove it from the PR, and if we don't I'll keep it there with some explanation why it wasn't implemented.

closes #2172

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Code is clear and sufficiently documented
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [x] ~End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design