Add bigint json parser (Issue #203, PR #2543)
# Description

This does not touch the handling of big numbers in forms.
So creating and editing service instances still needs work.
I have created a separate ticket for this as it seemed out of scope. (https://github.com/inmanta/web-console/issues/2542)
So this ticket focuses simply adding support for large integers coming from the api into the application.

closes #203

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Code is clear and sufficiently documented
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design