Fix result filter on Compile Reports page by handling boolean values (Issue #2496, PR #2633)
# Description

For the parsing and serialising of filter values:
I fixed this inline with how the special cases currently work.
But this needs a better way of handling these special values.
I created a ticket for this: https://github.com/inmanta/web-console/issues/2632

closes #2496

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Code is clear and sufficiently documented
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design