Fix empty api url by including origin in BaseUrlManager (Issue #2699, PR #2752)
# Description

I have changed the PrimaryBaseUrlManager to also use the origin in finding out the baseUrl.
This is only used for the API. Not for routing.
Routing now uses the `basePathname`.
Routing can't use the `baseUrl`, since the baseUrl contains the origin, and the origin can contain a port.
Routing would fail on the port because it thinks `:8888` is a path param.

closes #2699

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Code is clear and sufficiently documented
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design