Fix storybook errors & Improve small issues (Issue #2755, PR #2761)
# Description

I tried adding the `COMMITHASH` global to storybook as it was crashing a story.
I did not succeed 😅 . So I added the commitHash into the DependencyProvider allowing us to inject it.
Also fixed some other small issues I noticed.

closes #2755

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Code is clear and sufficiently documented
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design