Refactor CompareSelectionWidget (Issue #2877, PR #2893)
# Description

The original implementation was a ChipGroup. Increasing the fontsize broke the alignment a bit. So I just created it from scratch. Semantics wise, it never was a chip group. So I think it is fine for this to be a custom component.

closes #2877

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Code is clear and sufficiently documented
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design