Fix issue when the resource progress bar doesn't update automatically (Issue #3242, PR #3262)
# Description

The data for the progress bar is cached so it provides a more consistent look when paging.
In this case the table shows the loading view (as the state of the data is set to loading), but when the regular updates happen after 5 seconds, we don't show a loading view, only update the table with the new results.
The `kind` of the data is not updated in this case, so the whole update of the progress bar was skipped. With this PR, the contents of the response are taken into account as well for the update.

closes #3242

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Code is clear and sufficiently documented
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [x] ~End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design