Fix issue when the resource progress bar doesn't update automatically (Issue #3242, PR #3262)
The data for the progress bar is cached so it provides a more consistent look when paging.
In this case the table shows the loading view (as the state of the data is set to loading), but when the regular updates happen after 5 seconds, we don't show a loading view, only update the table with the new results.
The `kind` of the data is not updated in this case, so the whole update of the progress bar was skipped. With this PR, the contents of the response are taken into account as well for the update.