Ignore dependabot upgrades for react 18 types (PR #3308)
# Description

The types for react 18 contain some breaking changes. While updating our code is possible, the libraries we depend on haven't been upgraded yet either, so the type checking still fails. I think it's best if we wait until the major libraries we depend on (e.g. patternfly, easy-peasy) are upgraded to handle this correctly

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Code is clear and sufficiently documented
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [x] ~End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design