Remove outdated manual dependency resolutions (PR #3576)
# Description

The majority of the remaining ones are due to a transitive dependency on webpack 4 via storybook. When that's removed, we can remove the manual resolutions as well. (We already use webpack 5 for building the application itself, so no changes necessary there).

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] ~Attached issue to pull request~
- [x] Changelog entry
- [x] Code is clear and sufficiently documented
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [x] ~End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design