Scroll into view when new lines are being added to the report while it is compiling. (Issue #3855, PR #3998)
# Description

The CodeHighlighter component can now get the "scrollBottom" attribute set to scroll the text output down. It will update whenever new lines are being added. By default the component will not scroll down to the bottom of the code block.

The attribute has been added to the output stream section and the error stream section on the Compile Report detailed view.

closes : 3855 - compiler report for in-progress compile: follow output *https://app.zenhub.com/workspaces/frontend-600a82c1f666040010200918/issues/inmanta/web-console/3855*

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Code is clear and sufficiently documented
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- ~~[ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design