An update banner will be displayed if your application happens to be outdated. (Issue #3879, PR #4220)
# Description
From now on, a banner will be displayed whenever the cashed version of the application is outdated. The check is done based on a version file that is generated at each build and will contain the commitHash, the date and the version number from the package.json file. This results in a json file that is being retrieved from the server through an API call. If the local bundled commitHash is not the same, or if the file doesn't exist yet, the banner will be displayed, inviting the user to hard reload his browser.

<img width="1279" alt="image" src="https://user-images.githubusercontent.com/44098050/196184780-7b394276-d568-48dd-942a-15eff4a6d7f4.png">

closes *#3879*

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Code is clear and sufficiently documented
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ]~~End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design