Improve the sidebar closing behaviour when on smaller screens. You can now click in the page content to close the container. (Issue #4119, PR #4240)
# Description

Removed unneeded code and made the side-bar self managed (based on behaviour provided by Patternfly)
The result is that when you are on a small screen and the sidebar becomes overlay, you can easily close it by clicking anywhere in the page content. The hamburger icon is also still working to toggle the container.

closes *#4119*

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Code is clear and sufficiently documented
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
-~~ [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design