Bugfix for the scroll into view behaviour. (Issue #4259, PR #4263)
# Description

Adjust the previously implemented scroll into view for reports. Now, when the user does scroll upwards, the scroll to end of the container is being deactivated. I used the native dataset attribute to store the positioning (this doesn't cause re-renders as it isn't a state/prop change) and allows to track if the user scrolled upwards.

closes *#4359*

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Code is clear and sufficiently documented
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~~

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Code is clear and sufficiently documented
- [x] Correct, in line with design