Improve alignment of filter options on smaller screens. (Issue #4339, PR #4380)
# Description

Tried to find an approach to resolve the issue for all the filters, not only the ones on the Desired state page.  I would have personally removed the vertical spacers, but this is advised by the Patternfly guidelines when having multiple types of actions in your toolbar. This means that when the Compile button jumps on the second line, it won't be nicely aligned with the content above in some cases.

![filter-alignment-improvement](https://user-images.githubusercontent.com/44098050/204239063-c106fb73-f769-4997-8f44-a315c521f9a2.gif)

closes *#4339*

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Code is clear and sufficiently documented
- [x] Correct, in line with design

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design