Add the needed scripts for a clean setup for E2E automation. (Issue #4305, PR #4374)
# Description

This PR closes the first steps in the automation for the e2e tests. Two scripts have been added separately, and are being sequentially called in the jenkinsfile to run the tests.

Both commands can be ran independently from the pipeline.

## setup-server.sh
This will pull in the local-setup repository from Gitlab. Builds the docker containers for the orchestrator and install the lsm-test modules. It will always copy the latest dist available on your current branch (The jenkins pipeline is building a new version in the e2e stage). But it means you can have any version locally you wish to run the tests.

## clear-server.sh
This script is meant to be run AFTER you have setup a server. It will destroy the docker containers and remove the temp folder containing the local-setup repo.

closes *#4305*

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Code is clear and sufficiently documented

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design