Adding automated e2e testing for the Service Catalog, for a basic-service instance. (Issue #4317, PR #4395)
# Description

Added the first part of the Service Catalog scenario. I used a few methods to await on the compile to be finished. I suppose we can optimise them maybe, but we are tied to the asyncness of Cypress. While loops are not supported.

closes *#4317*

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Code is clear and sufficiently documented
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design