Fix jenkins leaks and misconfiguration of ports. (Issue #4386, PR #4407)
# Description

This PR relies on the changes done in the local-setup repo. The binding to the port in the docker-compose file have been adjusted to 127.0.0.1 instead of falling back on 0.0.0.0

Removing the ports from the yml file is not possible because of Mac and Windows OS not being able to work it out. Only works on linux.

closes *#4386*

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry

# Reviewer Checklist:

- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Code is clear and sufficiently documented
- [ ] Correct, in line with design