Fix cleanup issue in the build part 2 (Issue #4459, PR #4464)
# Description

Second attemps to fix the node-fetch issue. According to documentation, the LTS version of node supports it's own fetch, so we shouldn't need to import it anymore in the script file and also we could remove the dep from the package.json file.

closes *Add ticket reference here*

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [ ] Changelog entry
- [ ] Code is clear and sufficiently documented
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )