Cleanup scripts and repair tests that were broken by V2 update. (PR #4827)
Updated the scripts to repair the broken builds, since the upgrade to Yarn V2.  I also updated the readme file with the description of the commands, although, there are two of them I'm not sure what the purpose is. The relate to webpack profiles, but I left them open until we have time to figure them out.