Add support to run the e2e tests against the OSS Orchestrator releases. (Issue #4660, PR #4876)
# Description
I added the needed changes for the test suite to be compatible for the OSS version. The commands have been added in the package.json. Some tests are completely skipped, because they only apply to ISO.

I didn't add it to the jenkinsfile, since this will be setup in a nightly build. (the first scenario also succeeds for oss, my terminal truncated the output )

![image](https://github.com/inmanta/web-console/assets/44098050/6d937582-17e3-46e5-a741-a772454f5573)