UI-fix for the header when authentication is enabled. (Issue #4865, PR #4891)
# Description

Solves [#4865](https://github.com/inmanta/web-console/issues/4865)
#4871

This PR only resolves a part of the keycloak issues. The dropdown is now displayed neatly. There are still issues with the state management and the caching. Also the logout action doesn't work yet, but this will be tackled in the next ticket.

https://github.com/inmanta/web-console/assets/44098050/5f8bb434-4880-4578-a5ce-2dadef76a407