Page redirection has been fixed when the authentication token expires. (Issue #4885, PR #4902)
# Description

The redirection will be solved in the local-setup configuration since as of 18+ version of Keycloak server, we need a valid redirection URI set on the server-side.

Aside from that, the problem with the HEAD-calls are now resolved. They weren't sending the headers with the bearer, causing them to fail when Keycloak was enabled.

Lastly, there was still an issue when the token expired. The page would stay stuck and show a 403. It will now redirect to the login page.

The only issue remaining, is the state update in the profile dropdown, this will be resolved in a next ticket.

https://github.com/inmanta/web-console/assets/44098050/0e934a49-1bbb-4154-b74d-fff21de5b685