Improve the behavior on Firefox when hovering over code-block icons. (Issue #4916, PR #5172)
# Description

I tried to figure out which was the root cause of the problem, I changed the implementations a few times, removed elements but it's still not clear what caused the glitches. It only happens when the codeblocks contain a lot of information, impacting performance. Most of the glitching could be either related to the Tooltips (they have minimal glitches on the PF site on Firefox) combined with the styled components "´css" declaration. I removed any superficial wrapper/customization of elements.

It does still glitch once in a while when clicking on the kebab-menu and then clicking outside of it multiple times. This might be resolved once the deprecated version of the kebab menu is updated. But since I've been tackling that in another ticket, I kept it separated.

https://github.com/inmanta/web-console/assets/44098050/ea1dcce3-af45-4875-b0f9-923710ca350e