Add support for suggestions in the forms (Issue #5531, PR #5581)
# Description

The ServiceInstanceForm now handles suggested values, literal ones and parameters. The suggestions are both available in the Composer and Normal Form.

MSW has been added to the project to handle mocking API requests. [documentation](https://mswjs.io/docs/api/http-response)
React-Queries are now handled in a new folder labeled V2. The idea is to use this location for all queries we will add/edit.

There was also a bug resolution for the InlineInput Element, by adding typing, it turned out the event and value for that onChange event were inversed, causing the app to break when you were in expert mode and trying to update a dictList using a TextListInput.

closes *#5531*

https://github.com/inmanta/web-console/assets/44098050/3da1706f-3e39-4b7d-9645-50874757b498

https://github.com/inmanta/web-console/assets/44098050/2691d757-c569-494b-a9c2-65662ecae5b4