Added support to display attribute annotations in tabs in the Service Inventory (Issue #5532, PR #5613)
# Description

closes : #5532

The attribute annotations are now supported by the Service Inventory table.
When an attribute has the correct type of annotation, it will be rendered as a link button in the attributeTreeTable and upon being clicked on, it will open the matching tab containing the documentation.

Mermaid has been blocked from major updates and locked onto version 9.
The structuredClone method is not yet supported by JsDom and Jest. Adding the needed mocks would be a lot of work and counter productive. Version 9 of mermaid will still have all the needed security updates for now.

https://github.com/inmanta/web-console/assets/44098050/e639d996-9bd0-4fb7-b4c5-86404ef6b55d

https://github.com/inmanta/web-console/assets/44098050/5fe70936-7707-47f4-a314-71934c6f36af