resolve bug in duplicate form for preselected values in dropdowns. (Issue #5588, PR #5681)
# Description

I replaced the select as it was before, whilst keeping the original type of the options. Previously the values were stringified in the Object.keys mapping. Now that the values keep their typing, we can safely remove the logic in the mapping prepping the body to send to the api that was previously assuming the value would always be a string in case of enums. The issue wasn't specific to the duplicate form.

close #5676