Editing, Creating instances can now be done using a JSON editor. (Issue #5763, PR #5816)
# Description

Implementation for the React-Monaco-Editor.
The form now supports toggling between the normal form, and the JSON-Editor with linting.
The form-toggle is only enabled when the editor contains parseable and valid JSON according to the provided JSON-schema.
Only caveat is for integers or floats when the value is empty, it is displayed/parsed as an empty string. The value will then be considered invalid integer.

The Editor doesn't render properly in JSDOM, Hence the coverage will be done in the E2E tests.

closes #5763

https://github.com/inmanta/web-console/assets/44098050/54838fe8-7b34-4466-9da7-75c308d950ae