Allow the user to copy either the UUID or the identifier of an instance. (Issue #5764, PR #5822)
# Description

closes #5764

Created a standalone component that can take in multiple options/values that can each be copied to the clipboard. The service inventory row will use this copy button/dropdown when there is an identifier available.

Copy to clipboard can't accurately be tested, so I added a check in the E2E to see if there are two options available as expected.

https://github.com/inmanta/web-console/assets/44098050/cddf1144-8231-4f41-a84a-0030dfa07e03