Implement Instance Actions and Expert Mode on the instance details page. (Issue #5859, PR #5966)
# Description

Implementation of the different actions that can be performed on an instance, including the expert actions.
The attributesTab > JSON view will have a Force Update button in expert mode, and an expert action menu is available with the destroy action and force state action.

There is a known issue with the expert modal. When closing the modal, the dropdown of the expert action menu remains open.

Closes #5859