Implement the Events tab on the Instance details page. (Issue #5781, PR #6024)
# Description

**Feature:**
The Instance Details has been extended now to display the Events for a selected version in the history.
The user can now consult the events present in the history logs in a dedicated tab in the main section of the page.
The rows can be toggled and will be highlighted in a golden hue when an event is an errored transition.
In the history logs, the instance_version displayed in the events is the selected version minus 1. This has yet to be estimated to be an issue or not, since on the main event page, using the events endpoint, the data seems to be slightly different.

The Compile report link will display a different text depending on the type of report. It has been discussed with other teams, and for now, we can assume that any report containing the "validate/validation" keyword is a validation report, all others are export reports.

closes #5781

![image](https://github.com/user-attachments/assets/96dd300f-5a76-4156-8ebc-ff4a376f9873)