Improve the display of timestamps in the instance details page. (Issue #6047, PR #6049)
# Description

Updated the display of timestamps, and repaired the wrong attribute being used for the date display in the history section of the instance details page.
The headers of the tables are also updated to use the word timestamp instead of date.
The dateWithTooltip component has been extended with a new attribute to switch between the relative display, and full display of the passed date.

closes #6047

<img width="1043" alt="image" src="https://github.com/user-attachments/assets/b776bba1-48fb-454c-bbd7-253b05a3e9f8">