Hide Left sidebar from the Instance Composer when in the view mode (Issue #5988, PR #6081)
# Description

Hides Left sidebar from the instance composer viewer. it was done by hiding element, as conditional rendering was stopping lifecycle due to missing ref for sidebar that is required in further steps. It could be resolved but that would equal to some additional conditions, and display: none is straightforward, clean and simplest solution in my opinion

closes #5988

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [ ] Changelog entry
- [ ] Code is clear and sufficiently documented
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )