Adjusted the stencil/left sidebar elements to match design, fix issues with shadows on both sidebars (Issue #5987, PR #6082)
# Description

remove distance between stencil elements and removes shadows, adds borders to them.
Fix shadow in stencilSidebar not to be overflowed by canvas paper - I think that those shadows that align with the design are visually really good, fix is better that removal of them but can be discussed.

![Screenshot 2024-12-02 at 11 50 52](https://github.com/user-attachments/assets/b98e026c-3ddb-4514-983c-09a8b94653f7)
![Screenshot 2024-12-02 at 11 50 30](https://github.com/user-attachments/assets/338b6dac-b461-4587-8b13-f45611f4ca68)

closes #5987 #6038

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [ ] Changelog entry
- [ ] Code is clear and sufficiently documented
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )