Add support for emoji to the Documentation Tab (Issue #5931, PR #6115)
# Description

Add emoji support for Documentation Tab, it's unfortunately not outside of the box, there is option to append some emoji directly, not by unicode though. But that would require change of habits for user, and now we can handle what is used already.

![Screenshot 2025-01-07 at 09 44 14](https://github.com/user-attachments/assets/58d90502-03ec-47d8-9cfb-cbaf81528d27)

closes #5931

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [ ] Changelog entry
- [ ] Code is clear and sufficiently documented
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )