Implementation of a markdown previewer for the instance documentation. (Issue #6268, PR #6271)
# Description

Implement the Markdown Preview page.
A button has been added to the Instance Details page when in Expert mode.
The hint about this only being a preview tool can be closed.

Updated the e2e tests to be less prone to flake. Once in a while, the input field for the Find will not type all the characters correctly, causing unexpected search targets to be used instead of the correct one.

closes #6268

https://github.com/user-attachments/assets/dadfd3c2-cb45-46ad-a018-6075cc0b3706