Improve build checks in Jest to fail tests on console.error and console.warn. (Issue #6347, PR #6353)
# Description

Implement a tracker in Jest configuration for warnings and error messages. If any are present, they will be printed at the end of the tests.
The failure is on the full test suite that ran. This is to avoid masking the output of individual test cases.

closes #6347

preview of what the failure would look like when msw catches an unhandled http call:

![image](https://github.com/user-attachments/assets/bbc9f54f-e15e-4690-a6bb-36927cb45149)