Include the Json-bigint package in the project. (Issue #6379, PR #6382)
# Description

Remove json-bigint dependency and replace it with local implementation. The original package had tests written with Mocha, I made some adjustments to be able to run them with our Jest-unit tests. FIles I could move to Typescript, are also updated. The only two remaining are the original source-code for the two methods.

closes #6379