Revert the change to the desired state e2e test. (PR #6484)
# Description

Revert the changes that were needed because of the leak in the local-setup when booting an ISO7 orchestrator, it used more recent models, which don't report the lsm lifcycle transfers in the dry-runs. Now that the local-setup uses the correct ISO7 images and models, changes are reported again when performing a dry-run.